backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
python-decouple>=3.8
psycopg2-binary>=2.9
Pillow>=10.0
gunicorn>=21.2
whitenoise>=6.6
